+++ frontend/src/services/map/markerManager.js
@@ -1,49 +1,20 @@
-import markerGenerator from './markerGenerator';
-import placeholderImg from '../../assets/placeholder_128x128.png';
+import  markerService from './markerService';
+import fitterService from './fitterService';
 
-const parser = (item) => ({
-    id: item.id,
-    name: item.localization[0].name,
-    lng: item.longitude,
-    lat: item.latitude,
-    // TODO set place photo url
-    photoUrl: item.photoUrl || placeholderImg,
-    address: item.address
-});
-
-const removeMarker = (marker) => {
-    marker.remove();
-};
-
-const addMarker = (map) => (markerData) => {
-    const marker = wrappedCreateMarker(markerData);
-    marker.addTo(map);
-    return marker;
-};
-
-const createMarker = (markerData) => {
-    return markerGenerator.generateDefaultMarker(markerData);
-};
-
-const cacher = () => {
-    const cache = {};
-    return (funcToCache) => (data) => {
-        if (cache[data.id] !== undefined) {
-            return cache[data.id];
+const getService = (map) => {
+    let activeMarkers = [];
+    return {
+        setMarkersFromPlaces(...places){
+            markerService.clearMap(...activeMarkers);
+            activeMarkers = markerService.addMarkersFromPlaces(map,...places);
+            return activeMarkers;
+        },
+        setMarkersFromPlacesAndFit(...places){
+            const markers = this.setMarkersFromPlaces(...places);
+            fitterService.fitMap(map, ...markers);
+            return markers;
         }
-        const result = funcToCache(data);
-        cache[data.id] = result;
-        return result;
     };
 };
 
-const wrappedCreateMarker = cacher()(createMarker);
-
-export default {
-    clearMap(...markers) {
-        markers.forEach((marker) => removeMarker(marker));
-    },
-    addMarkersFromPlaces(map,...places) {
-        return places.map((item) => addMarker(map)(parser(item)));
-    }
-};
+export default {getService};
